Extract tests to owner file
