+++ spec/lib/rbs_infer/signatures/steep_bridge/lexical_scope_spec.rb
@@ -1,0 +1,6 @@
+require "spec_helper"
+require "rbs_infer"
+
+RSpec.describe RbsInfer::Signatures::SteepBridge::LexicalScope, :dummy_app do
+
+end
